+++ src/call/call.service.spec.ts
@@ -2,40 +2,14 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { CallService } from './call.service';
 import { CallRepository } from './call.repository';
 import { DataSource } from 'typeorm';
-
-import { newDb } from 'pg-mem';
-
-//TODO : 독립 setup파일로 빼기
-const setupDataSource = async () => {
-  const db = newDb({
-    autoCreateForeignKeyIndices: true,
-  });
-
-  db.public
-    .registerFunction({
-      name: 'current_database',
-      implementation: () => 'taxiDB',
-    })
-    .registerFunction({
-      name: 'version',
-      implementation: () => '1.0',
-    });
-
-  const ds: DataSource = await db.adapters.createTypeormDataSource({
-    type: 'postgres',
-    entities: [__dirname + '/../../src/**/*.entity{.ts,.js}'],
-  });
-  await ds.initialize();
-  await ds.synchronize();
-
-  return ds;
-};
+import { setupDataSource } from '../../test/util/setupDataSource';
 
 describe('CallService', () => {
   let service: CallService;
+  let dataSource: DataSource;
 
   beforeEach(async () => {
-    const dataSource = await setupDataSource();
+    dataSource = await setupDataSource();
     const module: TestingModule = await Test.createTestingModule({
       providers: [
         CallService,
